+++ qdl_klayout_extension/__init__.py
@@ -1,7 +1,12 @@
-from .constants import *
-from core import *
-from errors import *
-from examples import *
-from transformations import *
-from utils import *
-from visualize import *
+"""
+This is the main module of the library. It contains the main interface to the
+library, as well as the main classes of the library.
+"""
+
+from . import constants
+from . import core
+from . import errors
+from . import examples
+from . import transformations
+from . import utils
+from . import visualize