+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log
 
-- 4/22/203; v0.1.1.0: Created new shapes.
+- 4/24/2023; v0.1.2.0: Visualization updates.
+  - Restructured visualizations:
+    - Removed `plot_coords` and replaced with `plot_polygon` and `plot_path` that use `SimplePolygon` and `SimplePath`
+      objects instead of `CoordinatesList`s.
+    - Defined `get_xy_from_coords` to use across plotting methods.
+    - Defined `Line2DDataUnits` to `plot_path` easily.
+    - Modified `plot_pattern` to take as input a `Pattern` and a shape or a 2D array of shapes that muches the Pattern 
+      array.
+  - `Pattern.get_pattern_array()` now returns a 2D array of `Coordinates`, instead of a `CoordinatesList`.
+  - Updated `geometries` to return sorted `CoordinatesList`s. This fixes the warning issue you get with the definition
+    of a `CoordinatesList`, but creates a new issue with the actual `CoordinatesList` elements. There needs to be a
+    specified rotationally sorting angle, so that we can overcome this issue.
+  - Added transformation methods to `Shape`. 
+  - Updated examples to use new visualization methods.
+
+- 4/22/2023; v0.1.1.0: Created new shapes.
   - Added a `Shape` superclass for the existing `SimplePolygon` class and the new `SimplePath` class.
   - Added new geometries (circles, rings, arcs).
   - Moved line segment length calculation for `Line` within the class for consistency.